Renames __init__ to app
