+++ app/manage.py
@@ -1,7 +1,7 @@
 from flask_script import Manager
 from flask_migrate import Migrate, MigrateCommand
 
-from app.__init__ import app, db
+from app.app import app, db
 
 migrate = Migrate(app, db)
