+++ app/models.py
@@ -1,4 +1,4 @@
-from app.__init__ import db,app
+from app.app import db,app
 
 class User(db.Model):
     __tablename__ = 'users'
